refactor: Folder structure
