+++ src/Commands.cs
@@ -1,4 +1,5 @@
-﻿using CounterStrikeSharp.API.Core;
+﻿using Config;
+using CounterStrikeSharp.API.Core;
 using CounterStrikeSharp.API.Core.Attributes.Registration;
 using CounterStrikeSharp.API.Modules.Commands;
 using CounterStrikeSharp.API.Modules.Utils;
