+++ src/Config/SklepcsData.cs
@@ -1,0 +1,14 @@
+﻿using System.Text.Json.Serialization;
+
+public class SklepcsData
+{
+    [JsonPropertyName("WebsiteURL")]
+    public string WebsiteURL { get; set; } = "www.sklepcs.pl/yourshop";
+
+    [JsonPropertyName("ApiKey")]
+    public string ApiKey { get; set; } = "1234567890";
+
+
+    [JsonPropertyName("ServerTag")]
+    public string ServerTag { get; set; } = "server1";
+}
