+++ src/Config/SklepcsPermission.cs
