+++ src/SklepcsWebManager/SklepcsWebManager.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Net.Http;
 using System.Threading.Tasks;
 using CounterStrikeSharp.API;
+using Models;
 
 namespace SklepCSManager
 {
